🚚 move : AlarmCalculatorView 내에서 Cell 관련 파일 Cell 그룹으로 이동
