+++ Inception/Inception/Screens/AlarmCalculator/Cell/AwakeBasedRecoCell.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 class AwakeBasedRecoCell: UITableViewCell {
-
+  
   @IBOutlet weak var sleepCycle: UILabel!
   @IBOutlet weak var sleepLabel: UILabel!
   @IBOutlet weak var awakeLabel: UILabel!
@@ -17,14 +17,14 @@ class AwakeBasedRecoCell: UITableViewCell {
   @IBOutlet weak var addButton: UIButton!
   
   override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
+    super.awakeFromNib()
+    // Initialization code
+  }
+  
+  override func setSelected(_ selected: Bool, animated: Bool) {
+    super.setSelected(selected, animated: animated)
+    
+    // Configure the view for the selected state
+  }
+  
 }
